Сотрудники выделены в отдельный контроллер
