--- WebStore/Views/Home/Employee.cshtml
@@ -1,10 +1,0 @@
-﻿@model WebStore.Models.Employee
-
-
-<div class="container">
-    <h3>Сотрудник №@Model.Id</h3>
-    <p>@Model.FirstName</p>
-
-    <hr />
-    <a asp-controller="Home" asp-action="Employees">Сотрудники</a>
-</div>
