--- WebStore/Views/Home/Employees.cshtml
@@ -1,32 +1,0 @@
-﻿@model IEnumerable<WebStore.Models.Employee>
-
-<div class="container">
-    <table class="table table-bordered">
-        <thead>
-        <tr>
-            <th>id</th>
-            <th>Фамилия</th>
-            <th>Имя</th>
-            <th>Отчество</th>
-            <th>Возраст</th>
-            <th></th>
-        </tr>
-        </thead>
-        <tbody>
-        @foreach (var employee in Model)
-        {
-            <tr>
-                <td>@employee.Id</td>
-                <td>@employee.SurName</td>
-                <td>@employee.FirstName</td>
-                <td>@employee.Patronymic</td>
-                <td>@employee.Age</td>
-                <td>
-                    <a asp-controller="Home" asp-action="Employee" asp-route-Id="@employee.Id">Карточка</a>
-                </td>
-            </tr>
-        }
-        </tbody>
-    </table>
-</div>
-
