+++ WebStore/Views/Shared/Partial/_MainMenu.cshtml
@@ -20,6 +20,6 @@
         </li>
         <li><a asp-controller="Home" asp-action="Error404">404</a></li>
         <li><a asp-controller="Home" asp-action="ContactUs">Контакты</a></li>
-        <li><a asp-controller="Home" asp-action="Employees">Сотрудники</a></li>
+        <li><a asp-controller="Employees" asp-action="Index">Сотрудники</a></li>
     </ul>
 </div>
